pyproject.toml
[tool.poetry]
name = "elcortify-backend"
version = "0.1.0"
description = "Learning GraphQL from Django with elcortify"
authors = ["Rubén Solano <rubensoljim@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
django-model-utils = "^4.0.0"
django-environ = "^0.4.5"
psycopg2-binary = "^2.8.6"
graphene-django = "^2.13.0"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.3.5"
factory_boy = "^2.12.0"
django-test-plus = "^1.4.0"
pytest-django = "^3.8.0"
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
